Change constant polling as per Salami738 changes

+++ MintyBatteryMonitor.py
@@ -22,6 +22,7 @@ import pickle
 warning = 0
 status = 0
 debug = 0
+iconState = ""
 toggleFile = "/home/pi/MintyComboScript/Toggle.txt"
 PNGVIEWPATH = "/home/pi/MintyComboScript/Pngview/"
 ICONPATH = "/home/pi/MintyComboScript/icons"
@@ -43,16 +44,19 @@ def readData(filepath):
 
 
 def changeicon(percent):
-    i = 0
-    killid = 0
-    os.system(PNGVIEWPATH + "/pngview -b 0 -l 3000" + percent + " -x 650 -y 10 " + ICONPATH + "/battery" + percent + ".png &")
-    out = check_output("ps aux | grep pngview | awk '{ print $2 }'", shell=True)
-    nums = out.split('\n')
-    for num in nums:
-        i += 1
-        if i == 1:
-            killid = num
-            os.system("sudo kill " + killid)
+    global iconState
+    if iconState != percent:
+        iconState = percent
+        i = 0
+        killid = 0
+        os.system(PNGVIEWPATH + "/pngview -b 0 -l 3000" + percent + " -x 650 -y 10 " + ICONPATH + "/battery" + percent + ".png &")
+        out = check_output("ps aux | grep pngview | awk '{ print $2 }'", shell=True)
+        nums = out.split('\n')
+        for num in nums:
+            i += 1
+            if i == 1:
+                killid = num
+                os.system("sudo kill " + killid)
 
 
 def endProcess(signalnum=None, handler=None):
